RoleMethod trait created;
